+++ releasenotes.md
@@ -1,3 +1,6 @@
 Whats New in v12.2
 --------------------------
+- Fixed document upload for ingredients returning incorrect results
+- Update login method
+- Major re-write and improvements for the API
 - This release may include more changes, for full details please refer to the CHANGELOG